split SettingsNovopenSection into a separate view
