+++ iOS/Settings/AppSettings.swift
@@ -1,13 +1,8 @@
-import SwiftData
 import SwiftUI
 import Appearance
 
 struct AppSettings: View {
     @EnvironmentObject private var store: ValueStore
-#if canImport(CoreNFC)
-    @Environment(\.modelContext) private var modelContext
-    @Query(sort: \SavedPen.createdAt) private var savedPens: [SavedPen]
-#endif
     
     var body: some View {
         List {
@@ -16,35 +11,7 @@ struct AppSettings: View {
 #endif
             GlucoseUnitPicker()
 #if canImport(CoreNFC)
-            Section("NovoPen") {
-                if savedPens.isEmpty {
-                    Text("No saved pens yet")
-                        .secondary()
-                } else {
-                    ForEach(savedPens) { savedPen in
-                        NavigationLink {
-                            SavedPenEditorView(pen: savedPen)
-                        } label: {
-                            HStack {
-                                Text(savedPen.title)
-                                
-                                Spacer()
-                                
-                                Text(savedPen.insulinType.title)
-                                    .secondary()
-                            }
-                        }
-                    }
-                    .onDelete(perform: deletePens)
-                }
-                
-                Picker("Hide Airshots", selection: $store.airshotFilter) {
-                    ForEach(AirshotFilter.allCases) {
-                        Text($0.title)
-                            .tag($0)
-                    }
-                }
-            }
+            SettingsNovopenSection()
 #endif
             Section {
                 Toggle(isOn: $store.debugMode) {
@@ -64,16 +31,6 @@ struct AppSettings: View {
         }
         .navigationTitle("Settings")
     }
-    
-#if canImport(CoreNFC)
-    private func deletePens(at offsets: IndexSet) {
-        for offset in offsets {
-            modelContext.delete(savedPens[offset])
-        }
-        
-        try? modelContext.save()
-    }
-#endif
 }
 
 #Preview {
